move objects into the MokiRuby module
