+++ lib/moki_ruby/future_wrapper.rb
