+++ spec/spec_helper.rb
@@ -19,3 +19,5 @@ RSpec.configure do |config|
   config.order = :random
   Kernel.srand config.seed
 end
+
+include MokiRuby
